+++ tui/src/widgets/mem_reg_interface.rs
@@ -132,9 +132,9 @@ where
             O: OutputSource + ?Sized + 'a,
         {
             match data.bp.remove(&cur_addr) {
-                Some(val) => {data.sim.unset_breakpoint(val);},
+                Some(val) => {data.sim.unset_breakpoint(val as u8);},
                 None => {match data.sim.set_breakpoint(cur_addr) {
-                    Ok(val) => {data.bp.insert(cur_addr, val);},
+                    Ok(val) => {data.bp.insert(cur_addr, val as usize);},
                     Err(_e) => {},
                 }},
             };
@@ -147,9 +147,9 @@ where
             O: OutputSource + ?Sized + 'a,
         {
             match data.wp.remove(&cur_addr) {
-                Some(val) => {data.sim.unset_memory_watchpoint(val);},
+                Some(val) => {data.sim.unset_memory_watchpoint(val as u8);},
                 None => {match data.sim.set_memory_watchpoint(cur_addr) {
-                    Ok(val) => {data.wp.insert(cur_addr, val);},
+                    Ok(val) => {data.wp.insert(cur_addr, val as usize);},
                     Err(_e) => {},
                 }},
             };
